refactor(shims): remove four shim files and fix booking-service 'as never' cast
Four shims documented in #80 collapsed:

1. web/src/lib/booking-service.ts: 'as never' cast eliminated. Shared
   createBookingService now declares its dep type as
   `Pick<Api, 'bookings' | 'endpoints'>` (exported as BookingServiceApi)
   so the web wiring can pass `{ bookings, endpoints }` directly.

2. web/src/lib/messages.sv.ts: deleted. Single caller
   (use-submit-market.ts) now imports messageFor from @fyndstigen/shared.

3. web/src/lib/postgrest-utils.ts: deleted. safeFilterValue moved next
   to its single caller as web/src/app/search/sanitize.ts; the test
   was already at search/sanitize.test.ts and now imports locally.

4. supabase/functions/_shared/takeover-helpers.ts: deleted. The crypto
   re-exports (sha256Hex, generateCode, timingSafeEqualHex) are imported
   directly from @fyndstigen/shared/crypto.ts at the four call sites;
   the two flow constants (CODE_TTL_MS, MAX_CODE_ATTEMPTS) move to
   _shared/takeover-constants.ts (no re-exports, just the two values).

Closes #80
