+++ supabase/functions/_shared/takeover-constants.ts
@@ -1,0 +1,8 @@
+/**
+ * Constants used by the takeover edge functions.
+ *
+ * Crypto primitives are imported directly from `@fyndstigen/shared/crypto.ts`;
+ * this file holds only flow-specific tunables.
+ */
+export const CODE_TTL_MS = 15 * 60 * 1000 // 15 min
+export const MAX_CODE_ATTEMPTS = 5
