+++ web/src/hooks/market-form/use-submit-market.ts
@@ -4,7 +4,7 @@ import { useCallback, useRef, useState } from 'react'
 import { geo } from '@/lib/geo'
 import { runMarketMutation } from '@fyndstigen/shared'
 import { useDeps } from '@/providers/deps-provider'
-import { messageFor } from '@/lib/messages.sv'
+import { messageFor } from '@fyndstigen/shared'
 import type { MarketFields } from './use-market-fields'
 import type { OpeningHoursDraftResult } from './use-opening-hours-draft'
 import type { ImageDraftResult } from './use-image-draft'
