+++ web/src/lib/booking-service.ts
@@ -1,14 +1,11 @@
-import { createBookingService, createSupabaseBookings } from '@fyndstigen/shared'
+import { createBookingService, createSupabaseBookings, type BookingServiceApi } from '@fyndstigen/shared'
 import { supabase } from './supabase'
 import { endpoints } from './edge'
 
-// createBookingService expects an Api-shape object. We provide just the
-// two slices it actually uses (bookings + endpoints) and cast away the
-// rest. The cast is contained to this file.
-const bookingApi = {
+const bookingApi: BookingServiceApi = {
   bookings: createSupabaseBookings(supabase),
   endpoints,
-} as const
+}
 
-export const bookingService = createBookingService({ api: bookingApi as never })
+export const bookingService = createBookingService({ api: bookingApi })
 export type { BookingService } from '@fyndstigen/shared'
